fixedish player falling out of plane desynch bug
I am telling the server every 2 seconds where the seat is on the client side. thus when the server lags the seat will be within the player's tick distance when the server catches up so the client won't kick the player out of the seat entity. I documented my hour of hell on discord showing how I figured it out.