abandoning reform-passenger-system branch
the Forge PartEntity<?> class is really annoying. I don't know why I tried to force the rest of the mod to conform to its needs. literally broke everything. I should have realized much sooner that this was a garbage approach and wasted a lot of time.

of course the overall goal of making the player a direct passenger of a vehicle is still possible. although it would still be somewhat destructive. there has to be a better approach.