significant performance boost to vehicle layer texture rendering
before each layer texture was rendered as a separate model on top of the base model. this was horrendous. if there was 10 vehicles with 2 layers each, the game treats it as rendering 30 models.
now combines base and layer textures into 1 dynamic texture and only renders that 1 texture. this dynamic texture is generated once when the entity spawns and once per paint job update. now only one model is rendered per vehicle.